// examples/workflow/main/src/lib/model/enum/actions/mail-action-editor-params.enum.ts
/*
 * Copyright(c) RIB Software GmbH
 */

/**
 * Input/Output parameters for the action 'Mail Action'
 */
export enum MailActionEditorParams {
	From = 'From',
	Subject = 'Subject',
	Body = 'Body',
	Html = 'HTML',
	SaveMail = 'SaveMail',
	UseDefaultCredentials = 'UseDefaultCredentials',
	DocId = 'DocId',
	WebLink = 'WebLink',
	Receivers = 'Receivers',
	Cc = 'CC',
	Bcc = 'BCC',
	Attachements = 'Attachements',
	ReplyToList = 'ReplyToList',
	Report = 'Report'
}