// examples/workflow/main/src/lib/model/enum/action-editors/parameter-type.enum.ts
/*
 * Copyright(c) RIB Software GmbH
 */

/**
 * Defines if the action parameter is an input parameter or an output parameter.
 */
export enum ParameterType {
	/**
	 * Defines that an action parameter is an input parameter.
	 */
	Input = 'input',

	/**
	 * Defines that an action parameter is an output parameter.
	 */
	Output = 'output'
}